test: cleanup & comment observable tests
